+++ tests/formatter/input-output-pairs/inline-lambdas-w-comments.in.gd
@@ -1,0 +1,78 @@
+func foo():
+	# a
+	var x1 = func(): pass		# inline a
+	# b
+	var x2 = func(p1,p2,p3,p4,p5,p6,p7,p8,p9,p10,p11,p12,p13,p14,p15,p16,p17,p18,p19,p20,p21,p22,p23): pass # inline b
+	# c
+	var x3 = func baz(			# inline c
+		# d
+		p1,						# inline d
+		# e
+		p2,
+		p3,
+		p4,
+		p5,
+		p6,
+		p7,
+		p8,
+		p9,
+		p10,
+		p11,
+		p12,
+		p13,
+		p14,
+		p15,
+		p16,
+		p17,
+		p18,
+		p19,
+		p20,
+		p21,
+		p22,
+		p23						# inline e
+		# f
+	) -> int: pass ; pass ; var x := [ # inline f
+		# g
+		0,							   # inline g
+		# h
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0,
+		0						# inline h
+		# i
+	] ; pass					# inline i
+	# j