+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  - Added support for annotations in parser & formatter
  - Added support for unique node names in parser & formatter
  - Added support for property etters in parser & formatter
+ - Added support for inline-lambdas in parser & formatter
 
 ### Changed
  - Removed support for legacy (Godot `3.x`) GDScript from core testcases